requirements.txt
flask
openai
requests
numpy
faiss-cpu
dotenv